+++ packages/pi-subagents/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [18.1.0-nklisch.4] - 2026-08-27
+
+### Fixed
+
+- Prevent a blocking `get_subagent_result` call from also receiving the same background-agent completion as a delayed follow-up notification.
+
 ## [18.1.0-nklisch.3] - 2026-08-24
 
 ### Fixed